Stok Rehberi güçlendirildi. Yetkiler arttırıldı.

+++ lib/view/main_page/alt_sayfalar/stok/stok_liste/model/stok_listesi_model.dart
@@ -36,7 +36,7 @@ class StokListesiModel with NetworkManagerMixin {
   double? satisFiat4;
   double? satisKdv;
   int? alisDovTip;
-  int? bulunanFiyat;
+  double? bulunanFiyat;
   int? depoKodu;
   int? fiatBirimi;
   int? muhdetayKodu;
@@ -80,7 +80,6 @@ class StokListesiModel with NetworkManagerMixin {
   String? stokAdi;
   String? stokKodu;
   String? ureticiKodu;
-  
 
   static StokListesiModel? _instance;
